+++ pkg/sass-parser/lib/src/expression/from-props.ts
@@ -8,6 +8,7 @@ import {BinaryOperationExpression} from './binary-operation';
 import {BooleanExpression} from './boolean';
 import {ColorExpression} from './color';
 import {ListExpression} from './list';
+import {MapExpression} from './map';
 import {NumberExpression} from './number';
 import {StringExpression} from './string';
 
@@ -16,6 +17,7 @@ export function fromProps(props: ExpressionProps): Expression {
   if ('text' in props) return new StringExpression(props);
   if ('left' in props) return new BinaryOperationExpression(props);
   if ('separator' in props) return new ListExpression(props);
+  if ('nodes' in props) return new MapExpression(props);
   if ('value' in props) {
     if (typeof props.value === 'boolean') return new BooleanExpression(props);
     if (typeof props.value === 'number') return new NumberExpression(props);