+++ pkg/sass-parser/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add support for parsing list expressions.
 
+* Add support for parsing map expressions.
+
 ## 0.4.13
 
 * No user-visible changes.